replace int(...) with float(...) in places

+++ src/locpix/scripts/img_seg/classic_config.py
@@ -114,7 +114,7 @@ class InputWidget(QWidget):
         Args:
             config (dictionary) : Configuration dict"""
 
-        config["vis_threshold"] = int(self.vis_threshold.text())
+        config["vis_threshold"] = float(self.vis_threshold.text())
         config["vis_interpolate"] = self.vis_interpolation.selectedItems()[0].text()
         config["sum_chan"] = self.sum_chan.isChecked()
